feat(uiux): split up long View files and improve opening, closing of the detailed grammar descriptions
also try cleaning up some uiux when it comes to placeholder components
