+++ app-swift/fushigi/Components/TableView.swift
@@ -1,0 +1,65 @@
+//
+//  TableView.swift
+//  fushigi
+//
+//  Created by Tahoe Schrader on 2025/08/08.
+//
+
+import SwiftUI
+
+struct TableView: View {
+    let grammarPoints: [GrammarPoint]
+    @Binding var selectedGrammarID: GrammarPoint.ID?
+    @Binding var showingInspector: Bool
+    let isCompact: Bool
+
+    var body: some View {
+        Table(grammarPoints, selection: $selectedGrammarID) {
+            if isCompact {
+                TableColumn("場合") { point in
+                    VStack(alignment: .leading, spacing: 4) {
+                        VStack(alignment: .leading, spacing: 4) {
+                            Text(point.usage)
+                                .font(.body)
+                            Text(point.meaning)
+                                .font(.subheadline)
+                                .foregroundColor(.secondary)
+                        }
+                        .padding(6)
+                        .frame(maxWidth: .infinity, alignment: .leading)
+                        .background(Color.gray.opacity(0.2))
+                        .cornerRadius(6)
+                        .overlay(
+                            RoundedRectangle(cornerRadius: 6)
+                                .stroke(Color.primary.opacity(0.4), lineWidth: 1)
+                        )
+                        coloredTagsText(tags: point.tags)
+                            .font(.caption)
+                    }
+                    .padding(.vertical, 2)
+                }
+            } else {
+                TableColumn("場合") { point in
+                    Text(point.level)
+                }
+                TableColumn("使い方") { point in
+                    VStack(alignment: .leading) {
+                        Text(point.usage)
+                        Text(point.meaning)
+                            .font(.caption)
+                            .foregroundStyle(.secondary)
+                    }
+                    .lineLimit(nil)
+                }
+                TableColumn("タッグ") { point in
+                    coloredTagsText(tags: point.tags)
+                        .lineLimit(nil)
+                        .fixedSize(horizontal: false, vertical: true)
+                }
+            }
+        }
+        .onChange(of: selectedGrammarID) { _, newSelection in
+            showingInspector = newSelection != nil
+        }
+    }
+}
